backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
networkx==3.2
pandas>=2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
openai==1.3.0
google-generativeai==0.3.0
httpx>=0.25.0
